+++ app/routes/decks/$id/edit.tsx
@@ -9,8 +9,8 @@ import { useLoaderDataExtra } from "../../../utils/loader-utils";
 import { cls } from "../../../utils/misc";
 import type { PageHandle } from "../../../utils/page-handle";
 import { toast } from "../../../utils/toast-utils";
+import { DeckNavBarMenuComponent } from "./_ui";
 import type { LoaderData } from "./_utils.server";
-import { DeckNavBarMenuComponent } from "./index.utils";
 export { loader } from "./_utils.server";
 
 export const handle: PageHandle = {
