+++ app/routes/decks/$id/history.tsx
@@ -18,12 +18,9 @@ import {
 import { cls } from "../../../utils/misc";
 import type { PageHandle } from "../../../utils/page-handle";
 import { MiniPlayer } from "../../bookmarks/index.utils";
+import { DeckNavBarMenuComponent, QueueTypeIcon } from "./_ui";
+import { PRACTICE_ACTION_TYPE_TO_COLOR } from "./_utils";
 import type { LoaderData } from "./_utils.server";
-import {
-  DeckNavBarMenuComponent,
-  PRACTICE_ACTION_TYPE_TO_COLOR,
-  QueueTypeIcon,
-} from "./index.utils";
 
 export { loader } from "./_utils.server";
