+++ app/routes/decks/$id/index.tsx
@@ -17,12 +17,12 @@ import { cls } from "../../../utils/misc";
 import type { PageHandle } from "../../../utils/page-handle";
 import { MiniPlayer } from "../../bookmarks/index.utils";
 
-import type { LoaderData, PracticeEntryTableExtra } from "./index.server";
 import {
   DeckNavBarMenuComponent,
   QueueStatisticsComponent,
   QueueTypeIcon,
-} from "./index.utils";
+} from "./_ui";
+import type { LoaderData, PracticeEntryTableExtra } from "./index.server";
 export { loader } from "./index.server";
 
 export const handle: PageHandle = {
