+++ app/routes/decks/$id/practice.tsx
@@ -18,11 +18,8 @@ import {
 import { cls } from "../../../utils/misc";
 import type { PageHandle } from "../../../utils/page-handle";
 import { BookmarkEntryComponent } from "../../bookmarks/index.utils";
+import { DeckNavBarMenuComponent, QueueStatisticsComponent } from "./_ui";
 import type { LoaderData } from "./_utils.server";
-import {
-  DeckNavBarMenuComponent,
-  QueueStatisticsComponent,
-} from "./index.utils";
 
 export { loader } from "./_utils.server";
